+++ py/viaems/testcase.py
@@ -1,16 +1,20 @@
 import unittest
+import sys
 
-from viaems.connector import ViaemsWrapper
+from viaems.connector import SimConnector, HilConnector
 
 
 class TestCase(unittest.TestCase):
 
-    def assertWithin(self, val, lower, upper):
-        self.assertGreaterEqual(val, lower)
-        self.assertLessEqual(val, upper)
+    def assertWithin(self, val, lower, upper, msg=None):
+        self.assertGreaterEqual(val, lower, msg)
+        self.assertLessEqual(val, upper, msg)
 
     def setUp(self):
-        self.conn = ViaemsWrapper("obj/hosted/viaems")
+        if "--hil" in sys.argv:
+            self.conn = HilConnector()
+        else:
+            self.conn = SimConnector("obj/hosted/viaems")
 
     def tearDown(self):
         self.conn.kill()